Allow for filtering items in inventory
We now allow for the first word after `m!inventory` to be a tag, which
may then be followed by a page number. This tag will be fuzzy matched
against the tags for the items in the user's own inventory, and, if a
match is found, only items with that tag will be shown.